refactor: resource & map selectors
