+++ src/redux/selectors/utils.ts
@@ -1,0 +1,4 @@
+import {isEqual} from 'lodash';
+import {createSelectorCreator, defaultMemoize} from 'reselect';
+
+export const createDeepEqualSelector = createSelectorCreator(defaultMemoize, isEqual);
